Added the waypoint moving widget.  * Updated it to maintain relative command location when editing (good for custom formations)  * Added jumpjets to it.  * Does not support unit orders, as in guard.
Also makes it impossible to use ctrl-shift on a build order to make a wall around a building. The key to move commands could be changed. But if something isn't done I probably won't use it because I like to queue my squares and I have been fine for years of not moving commands around.

It would also be really nice if someone could make the entire custom-formation set of commands selected at once if one command is selected.

svn path=/trunk/mods/zk/; revision=728