Added grid energy requirement to turret firing system. For example now if the Annihilator has less than 50 energy connected to the grid that it is in it will show an energy icon over it's head. There is no energy drain, just a threshold.
The Annihilator will still fire though because it's script is in Stupid Dumb Bos. To implement fully read the value of the UnitRulesParam "lowpower" 1 = low power so prevent firing in aim weapon.

Why not just paralyse it you ask? Because paralyse is for something else you fools!! If you remove the para when it regains power I am going to toggle my solars to un-para it. If you don't remove and just let it decay you make it arbitrarily weak to emp.

svn path=/trunk/mods/zk/; revision=791