Added gunship hold position as it was removed in the engine a while ago and I really liked it. Basically it is exactly like a Tactical AI that you cannot turn off but coded in reverse. I removed strafe instead of adding it.
Exactly the same UI as tactical AI. If you set gunships to hold fire and it is out of orders it will not stray from it's position. If you give it an explicit attack order it will start strafing again.

The engine was changed such that units on hold position still automatically acquired targets, this is so that BA gunships (with no tolerance) can fire and S44 inf doesn't require a lot of facing micro. Unfortunately as soon as a gunship automatically acquires a target it automatically strafes.

svn path=/trunk/mods/zk/; revision=965