Increased Shield/Cloaker morphs 25s -> 30s
Fixed bug in morph that displayed incorrect morph time.

Fixed a few missing sounds because noone checks them after changing anything.

Increased mobile shield cost 480 -> 550 (so there is a bit of a reason to make static ones).

Nerfed mobile cloaker:
 * Cost 420 -> 600
 * Range 550 -> 440
 * Decloak Distance 100 -> 180

So now it costs resources to morph to mobile but it is still free to morph back.

Reduced height of missile hitvolumes.

svn path=/trunk/mods/zk/; revision=1404