Puppies lose their halos and grow horns, they go 200 elmos below the ground instead of 10^6 elmos into the air. This stops enemies from mournfully looking up into the sky when a Puppy they wanted to shoot fires.
A drawback or advantage of this system (depending on how you look at it) is you can select all your Puppies, even the airborne ones, by drag selecting the area they were in. The potential drawback is you can see their selection circles (but if it is a massive issue we could unitRulesParam it out in the selection widgets). 

They no longer lose their command queue when firing as that is just asking to be widgetted around. Maybe they should take a little damage when they grow new legs to prevent infinite Puppy Dirt Fountain. (Finite Puppy Dirt Fountain is perfectly ok with me)

Puppies also lose cloaking and stealth when taking a vacation because it is not needed underground and the cloak effect was sometimes visible when they returned. Cloak did not work anyway as there appears to be an inbuilt engine delay of cloak state change.

Puppies are immune to all damage while underground thanks to unitPreDamaged. If they attempt to seek refuge there they are kicked out within 15 seconds of entering.

Also:
 * Fixed rare error in tactical AI.
 * Fixed Cobra hitvolume
 * Added impulse to Sumo jump. It was always meant to have impulse but an engine bug in an old engine version caused it to fly off the map if it impulsed when it landed, something to do with colliding with the units it just caused to move.


svn path=/trunk/mods/zk/; revision=1463