Fixed stealth drawing properly.
Sythe and Sniper are no longer stealthed when not cloaked.

Area cloakers disable cloaking when turned off. This almost entirely prevents the existence of radar dots within enemy LOS. The only time this can now happen is in the time it takes for a cloaker to shrink, I would like to continue jamming during this time but that is a job for later.

svn path=/trunk/mods/zk/; revision=1830