Small Krow nerf instead of Licho's massive one that he tried to sneak through (he wants to do a nerf though so here is one). This should be it for nerfs, this one may even be too far.  * Speed 3.8 -> 3.55 (instead of 3.8 -> 3)  * Range 550 -> 480
Skuttle Damage 10k -> 8k. The damage may have been to high but I don't really know. There are few things with 8k to 10k HP so this does not change that much. It makes things that die in 2 shots easier to retreat and also protects Commanders that make it past 8k HP.

svn path=/trunk/mods/zk/; revision=2066