Rewrote paralysis using the Gordian Knot approach, this approach is really awesome. Has the same behaviour as previous EMP except without the morph clearing fix. Basically no computation is done in the gadget.
Rewrote the Sumo aimscript using MATHS. This method of writing aimscripts is far superior to the most commonly used methods; GUESSWORK and BLINDNESS. MATHS can be used to produce aimscripts with the very highly prized property of Not Being Utterly Retarded.

As such the aimscript works All The Time. Not just when heading = 270 or 90 and pitch <= 45.

svn path=/trunk/mods/zk/; revision=2346