Added mashup widget XrayHaloSelections. Not yet optimized WIP so not enabled.
This is a combination of Selection Xrays with SelectionHalo for cloaked only because lups cloak doesn't work with Xray shader.

Purpose: As opposed to shapes drawn around a unit's base, these selection effects look appropriate even if the unit is not on the ground (in flight or in transport).

To see ally selections go to "Settings > Interface > XHS - Show Ally Selections"

svn path=/trunk/mods/zk/; revision=3063