Reverted everything Kerr did since r3526.
r3530 change completely broke the Krow.
 * Dgun did not require manual firing
 * Gunflare and emit points completely screwed.
 * Turrets rotated incorrectly.
Basically this needed reverting, the whole thing may not be broken but something was and I cannot be bothered fixing it for you. I'm not letting bugs seep in for me to fix later.

While I was doing that I reverted Banshee as it's not much more work to revert more things:
 * Licho, Kingraptor, jK and I dislike it
 * The geometry is worse
 * Teamcolour is garish
 * Engines too black.
You can have another go at the banshee, the grey that occurs on the middle may be ok. Also no need to touch geometry, some faces were pointing out in the recent iteration.

svn path=/trunk/mods/zk/; revision=3531