Fixed commander inclusion in Initial States.
Firestorm:
 * Health 580 -> 660 (survives 1 Haxor)

Banisher:
 * Reverted Double ROF
 * Range 400 -> 460 instead

Major rework of Kodiachi main gun. Now shoots 3 shots in a spread with 80 damage each (down from 300). Panther should be the accurate raider of the factory. This hopefully helps Shieldbot although Kodiachi still outranges and outruns Bandit. Even if this doesn't make the unit less powerful I think it is a good direction to take the role.

svn path=/trunk/mods/zk/; revision=4950