Ensured Mex Snap and Area Mex prefer to build mexes at the same location. Added 'repair nanoframe' functionality to Mex Snap that gives a repair order to ally mex nanoframes at a location and blocks construction if the spot is already captured by an ally.
Sak wants to try slower gunships and I want to give them roles.

Brawler. It should emphasise range and not being hit by most land units.
 * Speed 4.7 -> 4
 * Uses more of it's range while strafing.

Blackdawn. Should be hit and run with ground interaction.
 * Speed 4.65 -> 4.5
 * Health 2900 -> 3400
 * Cruise Altitude 210 -> 150 (to compare Banshee is at 100)
 * Range 450 -> 300

Rapier. If it is going to defend other gunships it needs to be able to keep up with them and deal reasonable damage.
 * Speed 3.8 -> 4
 * Health 1300 -> 1100
 * Damage 150 -> 200 (1 shot better vs Vamp)
 * Range 300 -> 340

Banshee. This just needs to be a bit better.
 * DPS increased 5%
 * Range increased by 20

Gnat. Needs to be less extreme "completely win or die".
 * Health 90 -> 120 (This is over the Defender boundary, am I mad????)
 * Range increased by 20

svn path=/trunk/mods/zk/; revision=6134