Added new utility function GiveStateOrderToUnit().
To be used instead of Spring.GiveOrderToUnit() for state-changing commands such as CMD.FIRE_STATE etc.
It checks Spring.GetUnitStates() for the unit and doesn't call Spring.GiveOrderToUnit() when not necessary.
The reason is that according to all my tests Spring.GiveOrderToUnit() is VERY EXPENSIVE, approximately 30-50 times more expensive than Spring.GetUnitStates(). Tested with various state commands and in different circumstances and always got similar result.

svn path=/trunk/mods/zk/; revision=6533