With all the widget-toggling buttons in epicmenu, it was getting messy. It appears as though options are doubled ("Night View" and "Night View Config"). New setup:
If the widget has its own options, then make a submenu with the name of the widget. Inside that submenu, place button toggling the widget labeled with the word Toggle.

If the widget has no options, leave the toggle as it is, without being in a submenu.

Now when toggling a widget with options, it's instantly clear that it can be configured.

svn path=/trunk/mods/zk/; revision=9726