Dirtbag rework. It now leaves a wreck and a veh-passable hump. HP 550 -> 650, Sight distance 200 -> 400. Death explosion 60 AOE -> 128 AoE, 50 damage -> 150 damage.
This is to address the Dirtbags use as a CPU and map destroying hardcounter to veh/tank. It seriously nerfs them when spammed: You feed the enemy metal. It makes them block veh and bots equally, block shots better, but the counter with a constructor is much easier than mass restore (and large units crush them: actually not sure how to prevent this). To increase it's non-trolly use, the Dirtbag is a better scout and is capable of sniping out winds, fleas, or even 3 cloggers (60 metal) for a mex (50 metal). It also makes them chain slightly to further discourage spamming.

svn path=/trunk/mods/zk/; revision=9820