Lagmonitor: fix a case of unit from team 0 not given to other player when team 0 exit/resign/afk from game (because ALL spectator is on team 0 too, and this is unexpected case by lagmonitor gadget)
MexPlacement widget: fix CentralBuildAI(CBA) not working when areaMex + shift was issued to an even number of constructor (this is because MexPlacement notify the same order multiple time). Now it will only loop once.

Avoidance widget: fix a case of high ping cause user's order to be cancelled (this only happen to recent changes/not in stable).

svn path=/trunk/mods/zk/; revision=9861