Fixing hitvolume hax part 1: Mex, Solar, Fusion, Wind. Factories are next.
To see just how completely broken this currently is, put any (!) raider between the two towers of the fusion on the long axis: it is invulnerable.

A hitsphere at ground level MUST be no more than 8x it's footprint (on the shortest footprint axis!) otherwise a enemy unit can just walk into the hitvolume. You can raise the hitvolume off the ground to prevent fleas getting under it as easily, but a unit whose midpos is higher up and matches with this height can still possibly get inside it. Non-spherical hitvolumes use a SPHERE that encompasses the ENTIRE VOLUME when determining whether an ally can shoot at the target: This essentially means nonspherical hitvolumes are unusable for structures prone to hitvolume hax (Anything that cannot defend itself or move). Hitvolumes SHOULD NOT be made to cover every part of the geometry of a unit, but simply offer a solid centre of it smaller than the footprint to shoot at.

Also did some visual improvements to wrecks/death scripts and reduced the hitvolume of the mex wreck.

svn path=/trunk/mods/zk/; revision=9904