Fixed the Tacnuke Sub not firing at ground targets or at empty ground / sea bottom.
Fixing it required only moving the firepoint by -15 elmos in Z direction.
In Upspring it showed as moving it horizontally, but ingame it actually moved vertically up, above water.
Don't ask my why that way, something is bugged with the coordinates because moving it in Y direction in Upspring moved it horizontally ingame. :O

Anyway the bug had nothing to do with commandfire=true, it was exactly as bugged without the commandfire.

This commit fixes issue 1603. (finally!)

svn path=/trunk/mods/zk/; revision=10043