Duck back to tracking missiles. Reload time 3 -> 4, same DPS. This has synergy with the Archer, since the projectile can fire, will still hit a unit under impulse, and the high alpha works with Archer skirm factor and persistent DPS. Stats allow it to 2-shot a llt (suiciding) or to kill an LLT with archer assistance. The reload and alpha should (hopefully) allow raiders to waste it's shots, because it should require the Archer and not be anti-raider on it's own. May have to up the alpha further for this though.
Archer range 400 -> 300, impulse 40 ->30. First change is to allow it to screen for Ducks and Bouys, since previously it had too much range to really let either fire (Duck because it has less range, Bouy because it misses). The impulse is a general nerf because it's no longer the sole source of anti-unit DPS in a duck/archer pack (which was not viable to use anyway), and this should make it harder to deal with heavy units (which is meant to be the ducks role).

svn path=/trunk/mods/zk/; revision=11394