Startboxes: let AI start anywhere
Previously it couldn't pick a startpos at all and always got the default middle of startbox position.
Now it will be able to start anywhere (including out of the box for cheating AI). Bots that don't pick any startpos always get a default one in the middle of the map and thus will now start there (mostly relevant for Null AI).

The obvious scope of picking inside own box is not possible because AllowStartPosition does not discern bots (it gives all bots ID == 255) so it would not know which box to use.