Spawned commanders always face the middle of map
Previously they used to face the enemy box if there were just 2 teams.
With arbitrary boxes, the notion of facing the enemy box no longer
has to make sense. In most practical cases the behaviour is the same
since the enemy box is usually on the other side of the map.