Warrior → Ghoul; Tick → Wisp
Warrior is a somewhat boring OTA name which doesn't fit either of the Cloakybot themes. The unit has recently received fairly large changes so it seems like the best moment to do a rename.

Tick is there mostly because the other rename creates a good excuse to do more at the same time and the unit is fairly unused. The name is not OTA but almost: it seems to have been included in most popular unit packs. It also does not belong in either theme.

Both proposed names fit within the evil magic theme (like Gremlin, Spectre) and are really good at describing the units:
Ghoul - slowly walks with arms stretched forward and mauls everything that gets close.
Wisp - a volatile flash of light out of nowhere, which doesn't harm people directly but still leads to their demise.