Remove version number from Decloak Range widget
Since the name change is resetting user settings, remove the version
form the name. Do it fast so people is not yet used to this widget and
there are not many settings yet.

Also change "Decloak Area" to "Decloak Range" to be consistent with
filename and naming used in Settings.