Set duration beam laser lead limits such that they can all reliably kill a flea (the most agile unit). Fixes #1538.
Aircraft cannot accelerate as fast as Flea and tend to be larger so they are not a good test case. The lead limit needs to be as high as possible (while hitting Flea) to ensure that long and fast units will be targeted such that they spend as much time in the beam as possible. With the current settings and very good timing on the Stop command it is possible for a HLT to miss a Flea. However it can always hit a Flea running in circles so I consider the Stop case unimportant in the tradeoff.