-Major performance improvements: Now all the expensive calculations are staggered significantly, and it assigns one worker every 3 frames rather than 25 workers all at once every 45 or so frames. Both bandwidth and CPU usage are significantly more even as a result. Also as a side effect the latency from the time a worker finishes one job until they receive another is greatly reduced.
-Improved cost model:
Now the intelligent cost model prioritizes repairing porc as highly as for building small porc. This increases both porc and worker survivability.
The priority for expensive jobs has also been significantly reworked, so that it is flatter wrt increasing distance. It used to disfavor big projects a bit too much.

-Bugfixes: Switching to incremental assignment exacerbated a couple of latent bugs, which are now fixed.