Mex config: mex value override works on config mex
Previously the override only worked for mex spots created through engine metalmap detection.
Now it works with mexes provided through config (where the value was not provided). Previously value not provided would always be set to 2 regardless of what the override was.

This allows conveniently setting a standard mex size for the map. For example this config:

```
return {
    metalValueOverride = 1.75,
    spots = {
        {x = 123, z = 123},
        {x = 456, z = 456},
        {x = 789, z = 789},
    }
}
```

will now create mexes of value 1.75 whereas previously it would create mexes of value 2.