Make endgame stats toggleable, including mid-game
Added an action to toggle the endgame stats and awards display, bound by
default to Alt+F1. The hotkey can also call up the stats screen during
the game, which will display the stats and omit the awards panel. The
screen now includes a toggle button, which also displays the
currently-bound hotkey; this should allow players to discover the new
functionality.

Other notes and features:
. The stats graphs auto-update every 15 seconds, allowing you to see a
live feed during the game.
. The first time you view the stats, the first graph is displayed
immediately rather than showing a blank screen until you click one of
the buttons.
. The widget remembers which graph you were last viewing and displays it
immediately the next time you toggle the stats on again.
. The button for the currently-viewed graph is highlighted.
. Stats for teams you cannot view are not included in the graph; i.e. if
you are a player and view the stats mid-game the enemy stats will not be
shown (rather than showing a fake line at zero). Stats for all teams are
shown to specs, and to players when the game is over.
. When viewed mid-game, the Awards and Exit buttons are not shown.
. The screen can be toggled even after the game is over, so that you can
view the battlefield without having to reposition the endgame screen.
. The screen was already moveable and resizable. Now, though, if you
call up the screen mid-game and reposition it, when the game is over and
the post-game delay counter has elapsed the screen is repositioned to
its standard location so that it can still serve as an end-of-game
splash screen.
. The caption during mid-game is "Game in progress - Statistics". At the
end of the game, the caption reverts to the normal ones ("Victory!",
"Defeat!", etc).