Speed nerfs:  * Firewalker 2.1 -> 1.45  * Penetrator 2.4 -> 2.2
These changes seem opposed to Quants Rule, since both these artillery pieces are fairly fast. However, artillery should be slow in order to make positioning important, especially if the artillery is good against units. The Firewalker change is extreme but I think it originally had high speed because the unit sucked. When it stopped sucking the speed was overlooked.