Some Shieldbot changes to look at.
Racketeer:
 * Reload time 7s -> 10s.
 * Disarm time 9s -> 6s.
 * Damage 1500 -> 1800.
This is a significant nerf with the intention of specializing Racketeer against heavy units as it is currently a bit oppressive against units in the 150 to 400 cost range. Two Racketeers will be able to reliably lock down a heavy unit but an army of Racketeers will no longer be able to complete disable an army of medium-weight units.

Rogue:
 * Projectile speed 170 -> 200.
This is counter to Quant's Rule but sometimes this kind of change is required to see whether projectile speed is the solution to Rogues problems. What's the worst that could happen?

Felon:
 * Shield charge 1200 -> 1600.
Higher max capacity improves the rate at which Felon can suck charge out of the shield network. It also prevents it going from perfectly fine to dead against a Phantom. Now the Cloakbot player needs to wait do some damage or wait for it to engage before picking up massive value.