Fix #3433, Det tagged submerged at shoulder depth
Increased the modelradius (which controls when a unit is considered
submerged) to match the actual height, both visual and gameplay-wise
(the AA gun is at the very top - else it would be able to shoot "from
underwater" due to sticking out). Increased the hitvolume to match as
well.

This was a tiny gameplay nerf so to compensate, health 85800 → 86000.