Vandal damage 72 → 70
Affected shots to kill:
 * Raven, Thunderbird 14 → 15
 * Phoenix 10 → 11
 * Locust 12 → 13
Flight time 3s → 4s as a KR balance sort of compensation.

All bursty units are supposed to have round damage values (because these are easy to parse and multiply) and Vandal was the only one where rounding the displayed value but not the actual damage would introduce meaningful discrepancy.