Improve Knight unit AI when attacking
This is a QoL improvement for the Knight.  The Knight has issues with LoS because its weapon can't fire through other Knights.  When giving a group of Knights an attack command against a single unit, they would tend to bunch up and push each other forward, making the problem worse.

This change causes Knights to always use skirmish AI against ground targets.  The effect is that when Knights in the back start pushing forward, the Knights in the front move to the side.

skirmLeeway and skirmBlockedApproachFrames were also tweaked to make the Knight's positioning more aggressive, it is an assault unit after all.