Moved hoko's ancient COB fixer to the tools repo
No longer useful: the breakage was a one-off in 2016 and all our COB scripts have been fixed. I don't expect us to get any more COB scripts (or COB breakage) in the foreseeable future, but if we do the fix should be done in BOS. The fixer seems to have been introduced accidentally based on that commit's description, anyway.