Capture:  * Capture health is no longer based on cost.  * Capture damage is no longer multiplied by (3 - 2*(currentHealth/maxHealth)).  * Capture health is now currentHealth + 500.  * Dominatrix now deals 333 DPS instead of 78. This reflects our average HP/cost of around 3.5.  * Just like EMP and Disarm, accrued capture damage is scaled leave the capture progress of a unit unchanged when it receives damage.  * Capture damage is now affected by armour.
The upshot of the change is that Dominatrix is better against mid-to-high cost glass cannons and worse against low-to-mid cost assaults. The 500 bonus health gives Dominatrix a 1.5s minimum capture time to prevent it from instantly capturing small or low health units. This bonus health is intended to model some fixed mental resistance or firewall around each unit.

Ogre:
 * Damage 440 -> 420