Synchronize in-game ranks with lobby ranks
These are significantly lower resolution, but I can't seem to find
equivalent higher resolution sources elsewhere. These are the versions
used in the lobby, and appear to also be visually identical to those
used in the website.

Hopefully this should result in less confusion of the form 'silver is
higher than gold?'