Applied the next iteration of the 1.5x HP test.
Much of what was liked about 1.5x HP centred around raider-raider interactions and raider-economy interactions. The rest of the game was thrown a bit off, especially skirmishers and aircraft. Because of this, the next test is focused on raiders.
 * Removed the general 1.5x HP multiplier and the Lotus change.
 * Raiders have around 1.2x HP and 0.8x damage.
 * Unarmed land constructors and Wind Generators have 1.2x HP.
 * Mex costs 90 and has 750 HP.

The full changes are only applied to Cloaky, Shield, and Rover because alpha and/or high weight raiders are unlikely to be affected in the same way.