Blind fix for widget local include
include() within a widget seems to always use defaults in the 3rd arg,
and, given gdb says that VFS.Include sees the second, environment
parameter as a table that appears to be an entire widget, I believe
this is the culprit.

I cannot test this, as everything works fine when in a local skirmish.
Some engine code appears to relax some restrictions in devmode.
I cannot be sure that this is why, but it seems plausible.