Revert previous until widgetside reimplementation
XP can be fully widgeted (which is superior because
that automatically takes care of LoS haxes and makes
it so that only those who care pay the perf cost).

Keeping gadget XP for now until said widget exists so
that the people who care can transition seamlessly
without the XP-less period inbetween.

This reverts commit f779105d416d90c953529593df338639544bf0be.