GBC: Fix order spam, order cancellation
We were checking hash before we calculated it, therefore we kept sending
the same order to the same units to 'reassign' it, which spring
interpreted as cancelling the order.

In short, this meant builders would stop gormlessly every three
slow update processes. Higher options.updateRate.value would simply make
this stutter worse.

This doesn't seem to fix everything, but it helps.