Categorize constructors as economic units
The tooltip for the end graph implies both factories
and mobile constructors should show up as economy units.
Currently factories are categorized as Other, and constructors
as Army.

I assume this is related to changes to energy production and storage
in constructors, and the tooltip description is the desired behavior.