GBC: Remove another crash source
This happened whenever there was an unidentified enemy very, very
close to the location of a soon-to-be build site.

This could only really happen when:
 - you're building something on the frontline
 - you're using GBC to do so
 - you don't have clear los to the build site
 - you do have clear radar coverage of the build site

This meant it was both very rare, and typically happened in moments
where you really don't want to take time out to go bug hunting.

It's even harder to reproduce locally unless you knew what
you were looking for, since global spec control (for positioning
enemy units) was usually paired with globallos, which prevents
this condition from happening.