HACK: Mock up to see GBC jobs in grid
Do not merge!

This is done partly as a sanity test and is here so it is not lost.
Hook direction will be reversed later, and everything else will be
made less bad.

The plan is to eventually show allied GBC jobs through a minimal
receiver widget that is much less intrusive than GBC itself.