Economy Panel: Experimental color semantics
This tries to use green = good, red = bad.
Currently, green = positive number and red = negative number, whether
the situation is good or bad.

I'll use this for a while and see how it plays. If it doesn't end up
being terrible, I'll turn it into an option.
Otherwise, I'll remove it.