ButteryPlayerList: Duplicate from Crude
This should make it easier to compare between the three main
lists as we go - Crude, Deluxe, and this new list, which will
attempt to bring the features of Deluxe to the performance level
of Crude.

This name was very carefully considered, by picking a term that
was about the same distance from Crude as Crude was to Deluxe
in the dictionary.