GBC: Cancel jobs inverse to commands given
If we issue a command to reclaim a unit (or nanoframe), we probably
don't want to have jobs to repair it, resurrect it, or to build
an identical copy in its place.

If we issue a command to repair or resurrect a unit, we probably
don't want to have jobs to reclaim it.

Area commands are unaffected, as covering all of the cases that can
happen (such as those that involve alt permanent area commands)
is a headache.
Perhaps they shouldn't be.

This eliminates 4-7 seconds of fighting with your builder every time
you accidentally queue up a Desolator instead of a Caretaker.
(This was totally done for code correctness and not because
this was a regular occurrence, ahem.)