Remove 'experimental' widget tag
Our stables no longer contain "stable" in their name so this doesn't
actually prevent them outside stable. Only one widget actually had the tag
set to true and it was disabled anyway (some xray selection shader).
The game is mature enough to maintain stability all the time so this is
less because of the implementation not working and more of a design thing.