GBC: Remove in-widget area circle drawing
This is heavily bitrotten, and the documentation hasn't reflected
implementation for a long time.

Spring already draws a circle for us, as it knows it is an area command,
and it's very easy for spring and GBC to get out of sync over what
the active command is. Try right clicking, or using a keyboard shortcut.

This frequently leaves us in a state where GBC keeps drawing a bright
red/white circle for a command that is no longer active, and requires
that you enter the globalbuildcancel command again and perform it on
an empty area before you can make it go away.

This would be a fair bit of work to fix something that has little to no
value - and is perhaps even actively harmful if it teaches users that
this is what area commands look like - so lets remove it.