GBC: Unify code with UnitGained
This unifies UnitCreated, UnitFinished, UnitGiven, and the Initialize
code.

I expected there to be more of a difference given that UnitCreated
and UnitFinished get called twice, but apparently we really did
set allBuilders and call UpdateOneWorkerPathing twice.

Behaviour preserved for now I guess, though it's something to think
about later.

Mex -> territory behaviour is a mess. As far as I can tell, it
always has been. See the commentary in the UnitGone additions.
To ensure that every Gone always has exactly one Gained, we always
update the territory center only after the mex is complete.