Chili Scrollpanel: Fix child view test
self._inrtt is set only when the current control is being drawn,
but we care about children rendering themselves when they are
rendering their caches, too. Otherwise, anything that wasn't originally
in view will continue to be empty when we change scroll position, until
such as time as the children are invalidated separately.