GBC: State toggle consistency
The integral menu provides functionality to make sure state buttons
send the opposite of the state displayed, and make this the new,
consistent state over the entire selection.

GBC implemented its own handling, which caused the command to instead
independently invert the state of every unit in the selection.

This commit has GBC respect the state sent from the integral menu, and
makes the behaviour more consistent with all other state buttons.