GetWindowGeometry -> GetViewGeometry as needed
In most cases these are the same. However, when DualScreenMode is enabled,
these are quite different and almost always want positions relative to or
within the viewport.

These were used somewhat interchangeably.
For example, mission_galaxy_battle_handler had uses of both
GetWindowGeometry and GetViewGeometry for the same purpose.

This replaces all instances of GetWindowGeometry with GetViewGeometry.

While at first it seemed as if gui_chili_selections_and_cursortip did need
to retain references to GetWindowGeometry, since it has minimap logic,
it is unable to render within the minimap viewport in DualScreenMode and
so needs to know how to snap the tooltip within main viewport bounds.

When the results went unused, or the call was commented out, the call
has been removed.