Defence Range Indicator: Optimize
At 1Hz, every defensive structure on the map had its position's LOS
state checked, regardless of whether or not the unit would have always
been visible. This load becomes quite heavy in late teams games or FFA,
and was absolutely crippling as a spectator.

This allows early exits in a few more cases. LOS now is only checked if
the unit has previously left radar (and not returned), which is the case
for all allied porc, and the case for all porc as a fullview spectator.
Additionally, as a fullview spectator, we get all of the callins for
everyone, so we can completely remove the GameFrame callin until we
lose fullview.

We now also handle non-fullview spectators properly. Before, we
wouldn't get any of the callins for incomplete porc that suddenly entered
the spectator's (but not the spectated player's) view, so circles would
be missing until the porc completed and triggered the completed callin.
Completed porc would never gain a range circle, as no more callins would
trigger, unless they were captured.

Also get rid of a few drawlist leaks.