Styling and make command not interrupt current qeue
Attempt at fixing the syling (?) And also making the command not cancel the current qeue, makes it so that issuing an exclude command from an airplane on an airpad no longer excludes it even when the plane has left the pad. Which I don't think is a bad thing honestly.