Properly handle newly built caretakers.
Previously, the widget would queue commands while the caretaker was
being built. Then during the build the widget would see it's not doing
anything, and assume a user issued a command so the widget would ignore
the caretaker until it's idle. The result was that sometimes a freshly
built caretaker would reclaim all metal/energy within its range before
the widget took control of it.

Now we explicitly look for the UnitFinished event, at which point the
caretaker is ready to start working.