Update Hacksaw firing points
Fixes the firing points on hacksaw so missiles no longer come out of the center of it as well as changing hacksaw to .dae since s3o export is deprecated. Adds a laser pointer to it (For FW) which is hidden by unit script. I left it in in case anyone wanted to use it in the future.