Razor Rework/Update
Makes Razor look a bit better, instead of it being a hexagon trying it's hardest to be a circle it is now a cirlce, I've also touched up the doors to make their edges a little better as well as the inside circle. Finally under the guidance of ShadowFury I was able to improve the metallic/smoothness map on it as well. Note: If ao plate is not strong enough set decal size to 7, I just wasn't sure if that was too strong so I did 6. There are also a few new unittextures which were leftover from me messing around trying different things, I figured I'd leave them in case they could be used in the future. It also got a new buildpic for the slight difference. Death clone has not been updated however as I didn't see the need, I can do it sometime if it is more important.